+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - implement loading state for provider subscription detail overlay
 - Technical User Management
   - Enhance technical user table by adding StatusTag to the status column
+  - Enhance technical user table status column by adding new status 'pending deletion'
 
 ### Bugfixes